aio.core: Consolidate utils
