+++ aio.core/VERSION
@@ -1,1 +1,1 @@
-0.2.1-dev
+0.3.0
