+++ aio.core/aio/core/BUILD
@@ -11,6 +11,7 @@ pytooling_library(
         "functional/__init__.py",
         "functional/collections.py",
         "functional/decorators.py",
+        "functional/exceptions.py",
         "functional/generator.py",
         "functional/process.py",
         "functional/utils.py",
