+++ aio.core/aio/core/functional/__init__.py
@@ -4,7 +4,8 @@ from .collections import async_iterator, async_list, async_set
 from .decorators import async_property
 from .generator import AwaitableGenerator
 from .process import async_map
-from .utils import maybe_awaitable, maybe_coro
+from .utils import buffered, maybe_awaitable, maybe_coro, nested
+from . import exceptions, utils
 
 
 __all__ = (
@@ -14,5 +15,9 @@ __all__ = (
     "async_map",
     "async_set",
     "AwaitableGenerator",
+    "buffered",
+    "exceptions",
     "maybe_awaitable",
-    "maybe_coro")
+    "maybe_coro",
+    "nested",
+    "utils")
