+++ aio.core/aio/core/functional/exceptions.py
@@ -1,0 +1,4 @@
+
+
+class BufferUtilError(Exception):
+    pass
