+++ envoy.base.utils/envoy/base/utils/BUILD
@@ -3,6 +3,7 @@ pytooling_library(
     "envoy.base.utils",
     dependencies=[
         "//deps:abstracts",
+        "//deps:aio.core",
         "//deps:pyyaml",
         "//deps:pytz",
         "//deps:trycast",
