+++ envoy.code_format.python_check/envoy/code_format/python_check/BUILD
@@ -3,9 +3,9 @@ pytooling_library(
     "envoy.code_format.python_check",
     dependencies=[
         "//deps:abstracts",
-        "//deps:aio.core",
+        # "//deps:aio.core",
+        "//aio.core/aio/core/__init__.py",
         "//deps:aio.run.checker",
-        "//deps:envoy.base.utils",
     ],
     tags=["abstract-types"],
 )
