+++ envoy.code_format.python_check/setup.cfg
@@ -31,7 +31,6 @@ install_requires =
     abstracts>=0.0.12
     aio.core>=0.2.0
     aio.run.checker>=0.2.1
-    envoy.base.utils>=0.0.13
     flake8
     pep8-naming
     yapf
