+++ envoy.code_format.python_check/tests/BUILD
@@ -2,6 +2,9 @@
 pytooling_tests(
     "envoy.code_format.python_check",
     dependencies=[
-        "//deps:pytest-asyncio",
+        "//deps:abstracts",
+        # "//deps:aio.core",
+        "//aio.core/aio/core/__init__.py",
+        "//deps:aio.run.checker",
     ],
 )
